// client/src/components/AuthModal/AuthModal.styles.ts
import styled from 'styled-components';

export const Styled = {
    FormWrapper: styled.div`
        padding: 16px;
    `,
    InputWrapper: styled.div`
        display: flex;
        justify-content: space-between;
        width: 100%;
        padding-bottom: 16px;
    `,
    Label: styled.label`
        margin-right: 8px;
    `,
    Title: styled.h2`
        width: 100%;
        padding-bottom: 16px;
    `,
};